Welcome screen moved to /components folder and embedded inside a StackNavigator
