+++ App.js
@@ -1,58 +1,30 @@
-import React, { Component } from 'react';
-import {
-	Button,
-	Image,
-	Platform,
-	StyleSheet,
-	Text,
-	View
-} from 'react-native';
+import React from 'react';
+import { StyleSheet, View } from 'react-native';
+import { StackNavigator } from 'react-navigation';
 
-type Props = {};
-export default class App extends Component<Props> {
-	showUniversityPressed(event) {
-		console.log("Show university");
-	}
+import WelcomeScreen from './components/WelcomeScreen';
 
-	render() {
-		return (
-			<View style={styles.container}>
-				<Image source={require('./static/images/logo.png')} style={styles.logo} />
 
-				<Text style={styles.welcome}>
-					Welcome to UniNinja
-				</Text>
-
-				<Text style={styles.instructions}>
-					Press the button below to display the name of the first uni!
-				</Text>
-
-				<Button onPress={this.showUniversityPressed} title="Show University" color="#F5FCFF" />
-			</View>
-		);
-	}
-}
+const App = StackNavigator(
+    {
+        WelcomeScreen: { screen: WelcomeScreen },
+    },
+    {
+        headerMode: "float",
+        navigationOptions: {
+            headerTitle: "uni ninja",
+            headerTintColor: "#1a64db",
+        },
+    }
+);
 
 const styles = StyleSheet.create({
-	container: {
-		flex: 1,
-		alignItems: 'center',
-		backgroundColor: '#1a64db',
-	},
-	logo: {
-		marginTop: 60,
-		width: 60,
-		height: 60,
-	},
-	welcome: {
-		fontSize: 30,
-		textAlign: 'center',
-		color: '#F5FCFF',
-		margin: 30,
-	},
-	instructions: {
-		textAlign: 'center',
-		color: '#BBBBBB',
-		marginBottom: 10,
-	},
+    container: {
+        flex: 1,
+        backgroundColor: '#fff',
+        alignItems: 'center',
+        justifyContent: 'center',
+    },
 });
+
+export default App;
